Proof of Thom iso
We were talking at office hours on Tuesday about the proof in the notes of the general Thom isomorphism, in particular extending it to any MO-algebra E (I think Christian pointed out an error in the proof, and we didn't know how to fix it nicely).  I've typed in what I believe is a correct proof.  (This is from lecture 1)