Rewrote example 2.3.4, per discussion with Eric
Eric explained example 2.3.4 to me again.  I worked out the isomorphism $\eps^* \sheaf I (0)$ with $\sheaf I(0)/\sheaf I(0)^2$ based on Eric's suggestion in coordinates, but somebody really ought to check my work (and figure out how to do it in a coordinate-free way).  The issue of $E$ vs. $EP^0$ and $\L$ vs. $\L - 1$ is confusing (and related), and needs to be standardized in these notes.